requirements.txt
# PyTorch 라이브러리 (CPU 버전)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep Learning & Detection
ultralytics>=8.0.0
supervision>=0.16.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Dataset & Augmentation
roboflow>=1.1.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
Pillow>=10.0.0

# Tracking
lap>=0.4.0
motmetrics>=1.4.0
filterpy>=1.4.5

# Optional: For Web Interface
# streamlit>=1.28.0
# gradio>=4.0.0

# Optional: For Video Processing
# moviepy>=1.0.3
# ffmpeg-python>=0.2.0

# Development Tools
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0